Delete from database implemented

+++ consumer.py
@@ -9,6 +9,7 @@ def readFromBucket(requests):
 
     #TODO: dont need to get all objects, just get one, it will be the lowest key value
     #get_object requires a key, in order to know what the lowest key is, i need all the objects in the bucket
+    print(requests.list_objects())
     allRequests = requests.objects.all()
 
     size = sum(1 for _ in allRequests)
@@ -57,7 +58,8 @@ def getWidgetFromSQS(sqs, url):
     receipt_handle = message['ReceiptHandle']
     final = json.loads(message['Body'])
 
-    # sqs.delete_message(QueueUrl=queue_url,ReceiptHandle=receipt_handle)
+    sqs.delete_message(QueueUrl=queue_url,ReceiptHandle=receipt_handle)
+    logging.warning('Widget recieved from SQS')
     return (final[0],final[1])
 
 def writeToS3(s3, key, data, bucket):
@@ -77,30 +79,37 @@ def writeToDB(table, key, data):
     print("writing to database")
     logging.warning('item added to database')
 
-def deleteFromS3(whereTo, key):
-    obj = s3.Object(whereTo,key)
-    obj.delete()
-    # print(obj)
+def deleteFromS3(s3, whereTo, key):
+    bucket = s3.Bucket(whereTo)
+    allObj = bucket.objects.all()
+    for obj in allObj:
+        if obj.key.split('/')[2] == str(key):
+            obj.delete()
+            logging.warning('item deleted from S3')
 
-def delteFromDB():
-    print("deleting")
+def delteFromDB(table, data):
+    table.delete_item(Key = {'widgetId':data['widgetId'], 'owner':data['owner']})
+    logging.warning('item deleted from database')
 
 # CL syntax
 # {where to read from} {what type of read (sqs/bucket)} {where to write to (bucket/db)} {bucket to write to (if applicable)}
 
-logging.basicConfig(format='%(asctime)s %(message)s:', filename = "logs.txt")
+logging.basicConfig(format='%(asctime)s %(message)s:', filename =  "logs.txt")
 
 #use command line arguments
 s3 = boto3.resource('s3')
+client = boto3.client('s3')
 if len(sys.argv) <= 1:
     #if no arguments
-    # source = 'usu-cs5260-hackley-requests'
-    # requests = s3.Bucket(source)
+    source = 'usu-cs5260-hackley-requests'
+    requests = client.Bucket(source)
     whereTo = 'usu-cs5260-hackley-web'
-    storage = 1
-    sourceType = 0
-    sqs = boto3.client('sqs')
-    sqsUrl = 'https://sqs.us-east-1.amazonaws.com/912483513202/cs5260-requests'
+    storage = 0
+    sourceType = 1
+    dynamodb = boto3.resource('dynamodb')
+    table = dynamodb.Table('widgets')
+    # sqs = boto3.client('sqs')
+    # sqsUrl = 'https://sqs.us-east-1.amazonaws.com/912483513202/cs5260-requests'
 
 else :
     #get input type
@@ -112,7 +121,7 @@ else :
 
     if sys.argv[2] == 'bucket':
         source = sys.argv[1]
-        requests = s3.Bucket(source)
+        requests = client.Bucket(source)
         sourceType = 1
 
     if sys.argv[3] == "db":
@@ -125,8 +134,11 @@ else :
         storage = 1
 
 keepGoing = 0
+update = 0
+create = 0
+delete = 0
 
-while keepGoing < 5:
+while keepGoing < 1:
     #read the file and return the json object and key
     if sourceType == 1:
         info = readFromBucket(requests)
@@ -157,13 +169,15 @@ while keepGoing < 5:
             elif storage == 0:
                 writeToDB(table,key,data)
 
-        if request == "delete":
-            print("This was a delete request")
-            print("request " + request)
+        elif request == "delete":
             if storage == 1:
-                deleteFromS3(whereTo, key)
+                deleteFromS3(s3, whereTo, key)
             elif storage == 0:
-                delteFromDB()
+                delteFromDB(table, data)
 
-        if request == "change":
-            print("This was a change request")
+        elif request == "update":
+            print("This was an update request")
+
+        else :
+            print("this request was not recognized")
+            print(request)
